Matteo/meme/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
validators
python-Levenshtein